Vibify core architecture
