+++ crates/bpm_detection_core/Cargo.toml
@@ -4,25 +4,14 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-errors = { path = "../errors" }
-build = { path = "../build" }
-sync = { path = "../sync" }
 parameter = { path = "../parameter"}
 
 
 itertools = "0.14.0"
-midir = "0.11.0"
 derivative = "2.2.0"
-wmidi = "4.0.10"
 statrs = "0.18.0"
 serde = { version = "1.0.219", features = ["derive"] }
-log = "0.4.27"
-instant = { version = "0.1", features = [ "wasm-bindgen" ] }
 arraydeque = "0.5.1"
-
-[target.'cfg(target_os = "macos")'.dependencies]
-coremidi-hotplug-notification = "0.1.3"
-coremidi = { workspace = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 chrono = { version="0.4.41" }
