+++ crates/bpm_detection_midi/src/fake_midi_output.rs
