+++ crates/bpm_detection_midi/src/midi_output_trait.rs
