+++ crates/bpm_detection_midi/src/sysex.rs
