+++ crates/midi-bpm-detector-plugin/Cargo.toml
@@ -30,6 +30,7 @@ num-traits = "0.2.19"
 ringbuf = "0.4.8"
 serde = { version = "1.0.219", features = ["derive"] }
 mimalloc = "0.1.47"
+wmidi = "4.0.10"
 
 toml = "0.9.5"
